requirements.txt
amqp==5.1.1
appdirs==1.4.3
asgiref==3.2.10
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
billiard==3.6.3.0
boto3==1.24.61
botocore==1.27.61
cachetools==4.2.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clipboard==0.0.4
colorama==0.4.3
configparser==5.0.1
contextlib2==0.6.0
crayons==0.4.0
cryptography==37.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1
django-allauth==0.42.0
django-background-tasks==1.2.5
django-common-helpers==0.9.2
django-compat==1.0.15
django-cors-headers==3.4.0
django-cron==0.5.1
django-extensions==3.0.9
djangorestframework==3.11.1
google-api-core==2.8.2
google-api-python-client==2.58.0
google-api-python-client-stubs==1.11.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
instaloader==4.9.3
jmespath==0.10.0
jsonpickle==2.2.0
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
mypy-extensions==0.4.3
oauth2==1.9.0.post1
oauth2-client==1.2.1
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
parse==1.6.5
pathspec==0.9.0
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparser==1.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
selenium==4.4.3
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==4.1.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.10.1
typed-ast==1.5.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
wsproto==1.2.0
